v0.19.0 - revm 0.19.0

+++ src/lifecycle/receipt.rs
@@ -4,8 +4,8 @@ use revm::primitives::ExecutionResult;
 /// Create an Ethereum [`ReceiptEnvelope`] from an execution result.
 ///
 /// [`ReceiptEnvelope`]: alloy::consensus::ReceiptEnvelope
-pub fn ethereum_receipt(result: ExecutionResult, prior_gas_used: u128) -> ReceiptWithBloom {
-    let cumulative_gas_used = prior_gas_used.saturating_add(result.gas_used() as u128);
+pub fn ethereum_receipt(result: ExecutionResult, prior_gas_used: u64) -> ReceiptWithBloom {
+    let cumulative_gas_used = prior_gas_used.saturating_add(result.gas_used());
 
     Receipt { status: result.is_success().into(), cumulative_gas_used, logs: result.into_logs() }
         .into()
